wrote jewel code for ultra auto, split critical steps into separate files
